removed scripts and added console so that it can be ran from anywhere after being installed
